requirements.txt
# ComplyFormAI Backend Requirements
# Install with: pip install -r requirements.txt
# Python 3.13 compatible versions

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Testing & Development
requests==2.32.3
pytest==8.3.4
httpx==0.27.2

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.1